% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetic_retinopathy.R
\docType{data}
\name{diabetic_retinopathy}
\alias{diabetic_retinopathy}
\title{Real Data used in real data analysis for Bivariate Cure Rate Model}
\format{
A data frame with 10 variables:
\describe{
  \item{id}{Unique identifier for each patient}
  \item{obstime_trt}{Observed lifetime for the treated eye}
  \item{obstime_untrt}{Observed lifetime for the untreated eye}
  \item{risk_trt}{Risk indicator for the treated eye}
  \item{risk_untrt}{Risk indicator for the untreated eye}
  \item{status_trt}{Event(blindness) indicator for the treated eye (1 = event, 0 = censored)}
  \item{status_untrt}{Event(blindness) indicator for the untreated eye (1 = event, 0 = censored)}
  \item{laser}{Indicator variable for laser treatment}
  \item{type}{Categorical variable indicating type/classification}
  \item{age}{Age of the patient}
}
}
\usage{
diabetic_retinopathy
}
\description{
This dataset contains observed lifetimes(time to blindness), censoring indicators, and covariates
for diabetic retinopathy patients for treated and untreated eye.
}
\references{
Blair, A. L., Hadden, D. R., Weaver, J. A., Archer, D. B., Johnston, P. B.,
and Maguire, C. J. (1976).
\emph{The 5-year prognosis for vision in diabetes}.
\emph{American Journal of Ophthalmology}, 81(3), 383--396.
}
\keyword{datasets}
