% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{EM_func_exp}
\alias{EM_func_exp}
\title{EM Algorithm for Bivariate Cure Rate Model (Exponential  distribution as a special case of generalized gamma distributionin the shared frailty)}
\usage{
EM_func_exp(
  gamma10,
  gamma20,
  rho0,
  beta10,
  beta20,
  alpha10,
  alpha20,
  t1,
  t2,
  X,
  censor_type_vec,
  print_opt = FALSE,
  epsilon = 1e-04,
  conv_obsll = TRUE
)
}
\description{
This function executes the EM Algorithm as given in the paper, assuming  Exponential  distribution
as the special case of generalized gamma distribution, i.e \eqn{q=\sigma=1}
}
\details{
Refer to \code{\link{EM_func_qunfixed}} for full model specification  and parameter interpretation.
Note that the intial values of q and \eqn{\sigma} are not needed and user should give the initial values for only the rest of parameters
}
\seealso{
\code{\link{EM_func_qunfixed}} and paper for further details
}
