% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{EM_func_exp_mc}
\alias{EM_func_exp_mc}
\title{EM Algorithm for Bivariate Cure Rate Model (Exponential  distribution as a special case of generalized gamma distributionin the shared frailty)
Monte Carlo summation approach for approximating integrals}
\usage{
EM_func_exp_mc(
  gamma10,
  gamma20,
  rho0,
  beta10,
  beta20,
  alpha10,
  alpha20,
  t1,
  t2,
  X,
  censor_type_vec,
  print_opt = FALSE,
  epsilon = 1e-04,
  conv_obsll = TRUE,
  n_mc = 1000
)
}
\arguments{
\item{n_mc}{Sample size of the Monte Carlo sample used  in monte carlo summation for approximation of integrals}
}
\description{
This function performs the same procedure as \code{\link{EM_func_exp}}, with the
difference that the integrals involved in the EM algorithm are approximated using
the Monte Carlo summation approach.
}
\details{
The model assumes an Exponential distribution as a special case of the generalized
gamma distribution, i.e., \eqn{q = \sigma = 1}.


Refer to \code{\link{EM_func_exp}} for full model specification and parameter
interpretation. Monte Carlo summation is used to approximate integrals involved
}
\seealso{
\code{\link{EM_func_exp}}, \code{\link{EM_func_qunfixed}}
}
