% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{EM_func_gamma}
\alias{EM_func_gamma}
\title{EM Algorithm for Bivariate Cure Rate Model (Gamma distribution as a special case of generalized gamma in the shared frailty)}
\usage{
EM_func_gamma(
  gamma10,
  gamma20,
  rho0,
  beta10,
  beta20,
  alpha10,
  alpha20,
  sigma0,
  t1,
  t2,
  X,
  censor_type_vec,
  print_opt = FALSE,
  epsilon = 1e-04,
  conv_obsll = TRUE,
  method1 = "Nelder-Mead",
  method2 = "Nelder-Mead"
)
}
\description{
This function executes the EM Algorithm as given in the paper, assuming  gamma  distribution
as the special case of generalized gamma distribution, i.e \eqn{q=\sigma}
}
\details{
Refer to \code{\link{EM_func_qunfixed}} for full model specification  and parameter interpretation.
Note that the user should give the initial value of q is not needed and it is assumed to be equal to sigma0 as \eqn{q=\sigma}
}
\seealso{
\code{\link{EM_func_qunfixed}} and paper for further details
}
