% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{EM_func_gamma_mc}
\alias{EM_func_gamma_mc}
\title{EM Algorithm for Bivariate Cure Rate Model (Gamma distribution as a special case of generalized gamma in the shared frailty with Monte Carlo approximation)}
\usage{
EM_func_gamma_mc(
  gamma10,
  gamma20,
  rho0,
  beta10,
  beta20,
  alpha10,
  alpha20,
  sigma,
  t1,
  t2,
  X,
  censor_type_vec,
  print_opt = FALSE,
  epsilon = 1e-04,
  conv_obsll = TRUE,
  n_mc = 1000
)
}
\arguments{
\item{n_mc}{Sample size of the Monte Carlo sample used for approximation of integrals}
}
\description{
This function performs the same procedure as \code{\link{EM_func_gamma}}, with the
difference that the integrals involved are approximated using
the Monte Carlo summation approach.
}
\details{
The model assumes a gamma frailty distribution as a special case of the generalized
gamma distribution, i.e., \eqn{q = \sigma} , integrals are approximated using
the Monte Carlo summation approach.


Refer to \code{\link{EM_func_gamma}} for full model specification and parameter
interpretation. Refer to the paper for further details.
}
\seealso{
\code{\link{EM_func_gamma}}, \code{\link{EM_func_qunfixed}}
}
