% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{EM_func_qunfixed}
\alias{EM_func_qunfixed}
\title{EM Algorithm for Bivariate Cure Rate Model}
\usage{
EM_func_qunfixed(
  gamma10,
  gamma20,
  rho0,
  beta1_init,
  beta2_init,
  alpha1_init,
  alpha2_init,
  q0,
  sigma0,
  t1,
  t2,
  X,
  censor_type_vec,
  print_opt = FALSE,
  epsilon = 1e-04,
  conv_obsll = TRUE,
  method1 = "Nelder-Mead",
  method2 = "Nelder-Mead"
)
}
\arguments{
\item{gamma10}{Initial value of parameter vector \eqn{\gamma_1}}

\item{gamma20}{Initial value of parameter vector \eqn{\gamma_2}}

\item{rho0}{Initial value of parameter \eqn{\rho}}

\item{beta1_init}{Initial value of parameter vector \eqn{\beta_1}}

\item{beta2_init}{Initial value of parameter vector \eqn{\beta_2}}

\item{alpha1_init}{Initial value of parameter \eqn{\alpha_1}}

\item{alpha2_init}{Initial value of parameter \eqn{\alpha_2}}

\item{q0}{Initial value of parameter \eqn{q}}

\item{sigma0}{Initial value of parameter \eqn{\sigma}}

\item{t1}{Observed lifetimes for event 1}

\item{t2}{Observed lifetimes for event 2}

\item{X}{Covariate matrix}

\item{censor_type_vec}{Censoring pattern indicator:
\describe{
  \item{1}{Both events observed}
  \item{2}{Event 1 observed, Event 2 censored}
  \item{3}{Event 1 censored, Event 2 observed}
  \item{4}{Both events censored}
}}

\item{print_opt}{Logical; if TRUE, prints iteration details}

\item{epsilon}{Convergence tolerance}

\item{conv_obsll}{Logical; if TRUE, convergence is based on observed log-likelihood}

\item{method1}{Optimization method in M-step for \eqn{l_{c_1}} (cure component)}

\item{method2}{Optimization method in M-step for \eqn{l_{c_2}} (lifetime component)}
}
\value{
A named numeric vector containing estimated parameters and its observed log-likelihood:
\describe{
  \item{gamma10, gamma11}{Estimated \eqn{\gamma_1}}
  \item{gamma20, gamma21}{Estimated \eqn{\gamma_2}}
  \item{rho}{Estimated \eqn{\rho}}
  \item{beta10, beta11}{Estimated \eqn{\beta_1}}
  \item{beta20, beta21}{Estimated \eqn{\beta_2}}
  \item{alpha1, alpha2}{Estimated  \eqn{\alpha_1}, \eqn{\alpha_1} }
  \item{q}{Estimated \eqn{q}}
  \item{sigma}{Estimated \eqn{sigma}}
  \item{Obs_LL}{Final observed log-likelihood}
}
}
\description{
Estimates model parameters using the Expectation-Maximization (EM) algorithm
for the bivariate cure rate model. Refer to the paper for further details
}
\details{
The model assumes:

\itemize{
  \item \eqn{\gamma_1, \gamma_2, \rho} are associated with probabilities of cure combinations
  \item \eqn{\beta_1, \beta_2} are associated with scale parameters of the Weibull distribution
  \item \eqn{\alpha_1, \alpha_2} are shape parameters of the Weibull distribution
  \item \eqn{q, \sigma} are parameters of the generalized gamma frailty distribution
}

The EM algorithm iteratively performs:
\itemize{
  \item E-step: Computes conditional expectations
  \item M-step: Updates parameters using by optimizing \eqn{l_{c_1}} and \eqn{l_{c_2}} seperately wrt  to the respective relevant parameters
}

Convergence is determined either by:
\itemize{
  \item Change in observed log-likelihood, or
  \item Relative change in parameter estimates
}
}
