% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{EM_qfixed}
\alias{EM_qfixed}
\title{EM Algorithm for Bivariate Cure Rate Model (Fixed q)}
\usage{
EM_qfixed(
  gamma10,
  gamma20,
  rho0,
  beta10,
  beta20,
  alpha10,
  alpha20,
  q,
  sigma0,
  t1,
  t2,
  X,
  censor_type_vec,
  print_opt = FALSE,
  epsilon = 1e-04,
  conv_obsll = TRUE,
  method1 = "Nelder-Mead",
  method2 = "Nelder-Mead"
)
}
\description{
This function is similar to \code{\link{EM_func_qunfixed}}, except that
the parameter \eqn{q} is assumed to be fixed and optimization is done
w.r.t rest of the parameters using EM Algorithm,
this function is mainly used for profile likelihood approach where it is run for grid of
values for  \eqn{q} and the MLE is obtained for that value of \eqn{q} at which the final
 observed log-likelihood is maximum, see paper for further details
}
\details{
Refer to \code{\link{EM_func_qunfixed}} for full model specification
and parameter interpretation.
}
\seealso{
\code{\link{EM_func_qunfixed}}
}
