% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{EM_qfixed_mc}
\alias{EM_qfixed_mc}
\title{EM Algorithm for Bivariate Cure Rate Model (Fixed q) using monte carlo summation approximation for integrals}
\usage{
EM_qfixed_mc(
  gamma10,
  gamma20,
  rho0,
  beta10,
  beta20,
  alpha10,
  alpha20,
  q,
  sigma0,
  t1,
  t2,
  X,
  censor_type_vec,
  print_opt = FALSE,
  epsilon = 1e-04,
  conv_obsll = TRUE,
  n_mc = 10^3
)
}
\arguments{
\item{n_mc}{sample size of the monte carlo sample used for approximaton of integrals involved using monte carlo summation}
}
\description{
This function performs the same thing as \code{\link{EM_qfixed}}, only difference being that all integrals are approximated by monte carlo summation approach, refer to the paper for more details.
}
\details{
Refer to \code{\link{EM_qfixed}} for full model specification
and parameter interpretation.
}
\seealso{
\code{\link{EM_func_qunfixed}}
}
