% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppfunctions_doc.R
\name{Observed_ll_cpp}
\alias{Observed_ll_cpp}
\title{Observed Log-Likelihood for Bivariate Cure Rate Model}
\arguments{
\item{gamma1}{Parameter vector \eqn{\gamma_1}}

\item{gamma2}{Parameter vector \eqn{\gamma_2}}

\item{rho}{Parameter \eqn{\rho}}

\item{beta1}{Parameter vector \eqn{\beta_1}}

\item{beta2}{Parameter vector \eqn{\beta_2}}

\item{alpha1}{Parameter \eqn{\alpha_1}}

\item{alpha2}{Parameter \eqn{\alpha_2}}

\item{q}{Parameter \eqn{q}}

\item{sigma}{Parameter \eqn{\sigma}}

\item{t1}{Observed lifetimes for event 1}

\item{t2}{Observed lifetimes for event 2}

\item{X}{Covariate matrix}

\item{censor_type_vec}{Censoring pattern indicator
\describe{
  \item{1}{Both events observed}
  \item{2}{Event 1 observed, Event 2 censored}
  \item{3}{Event 1 censored, Event 2 observed}
  \item{4}{Both events censored}
}}
}
\value{
Numeric value representing the observed log-likelihood
}
\description{
Computes the observed log-likelihood of the bivariate cure rate model
for given parameter values and observed data.
}
\details{
Refer to paper for observed log-likelihood form for the biavraite cure rate model, and further details.
}
