% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data for Bivariate Cure Rate Model}
\usage{
simulate_data(
  x_sim,
  gamma1,
  gamma2,
  rho,
  beta1,
  beta2,
  alpha1,
  alpha2,
  q,
  sigma,
  lambda,
  lambda_c
)
}
\arguments{
\item{x_sim}{Input covariate vector}

\item{gamma1}{Parameter vector \eqn{\gamma_1}}

\item{gamma2}{Parameter vector \eqn{\gamma_2}}

\item{rho}{Correlation parameter \eqn{\rho}}

\item{beta1}{Parameter vector  \eqn{\beta_1}}

\item{beta2}{Parameter vector  \eqn{\beta_2}}

\item{alpha1}{Parameter \eqn{\alpha_1}}

\item{alpha2}{Parameter \eqn{\alpha_2}}

\item{q}{Parameter \eqn{q}}

\item{sigma}{Parameter \eqn{\sigma}}

\item{lambda}{Parameter \eqn{\lambda}}

\item{lambda_c}{Parameter \eqn{\kappa}}
}
\value{
simulated dataset as a dataframe containing all relevant columns as in observed data
\describe{
  \item{t1}{Observed lifetime for event 1}
  \item{t2}{Observed lifetimes for event 2}
  \item{obsvd_status_t1}{Censoring indicators for event 1 (1 = event, 0 = censored)}
  \item{obsvd_status_t2}{Censoring indicators for event 2 (1 = event, 0 = censored)}
  \item{cure_status_t1}{Cure indicators for event 1 (1 = cured, 0 = susceptible)}
  \item{cure_status_t2}{Cure indicators for event 2 (1 = cured, 0 = susceptible)}
  \item{x}{Covariate vector}
  \item{censor_ind}{Censoring pattern indicators combination of both censoring indicators:
    \describe{
      \item{1}{Both events observed}
      \item{2}{Event 1 observed, Event 2 censored}
      \item{3}{Event 1 censored, Event 2 observed}
      \item{4}{Both events censored}
    }
  }
}
}
\description{
Generates simulated data based on model parameters.
}
\details{
Refer to the paper, model assumed for more details on these parameters

\eqn{\gamma_1, \gamma_2, \rho} is associated with probabilties of cure combinations

 \eqn{\beta_1, \beta_2} are associated with scale parameter   and \eqn{\alpha_1, \alpha_2} are shape parameter in Weibull distribution for conditional lifetime of susceptible individuals

 \eqn{q, \sigma, \lambda} in generalized gamma distribution

  lambda_c - \eqn{\kappa}. decides censoring proportion, rate parameter in exponential distribution assumed for censoring time distribution
}
